+++ src/main/java/nextstep/courses/domain/session/policy/SessionStatus.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.session.property;
+package nextstep.courses.domain.session.policy;
 
 public enum SessionStatus {
     PREPARING, ENROLLING, CLOSED;
