+++ src/main/java/nextstep/payments/domain/PaymentRepository.java
@@ -1,7 +1,5 @@
 package nextstep.payments.domain;
 
-import nextstep.courses.domain.Course;
-
 import java.util.List;
 
 public interface PaymentRepository {
